Managing Mass Spectrometry Experiments

The `MsExperiment` class allows the storage, management and processing of all aspects related to a complete proteomics or metabolomics mass spectrometry experiment. This includes experimantal design, raw mass spectromtry data as spectra and chromatograms, quantitative features, and identification data.

For details, see https://rformassspectrometry.github.io/MsExperiment

This package is part of the RforMassSpectrometry initiative: https://www.rformassspectrometry.org/

MsExperiment()

# S4 method for MsExperiment
show(object)

experimentFiles(object)

experimentFiles(object) <- value

# S4 method for MsExperiment
spectra(object)

spectra(object) <- value

Arguments

object
An instance of class `MsExperiment`

value
An object of the appropriate class for the slot to be populated.

Details

- Data files stored in a [MsExperimentFiles()] object.

- Metadata: ...

- Mass spectrometry data, i.e. spectra and their metadata are stored as `Spectra()` objects.

- Chromatographic data will be stored as `Chromatograms()` objects.

- Proteomics identification data, i.e peptide-spectrum matches is defined as `PSM()` objects. They are generally joined with the spectra data.

- Quantification data is stored as `QFeatures()` objects.

Author

Laurent Gatto

Examples

## An empty MsExperiment object
msexp <- MsExperiment()
msexp

#> Object of class MsExperiment 

example(MsExperimentFiles)

#> 
#> MsExpF> fls <- MsExperimentFiles(mzmls = c("/path/to/f1.mzML", "/path/to/f2.mzML"),
#> MsExpF+                          mzids = "/another/path/to/id1.mzid",
#> MsExpF+                          fasta = "file.fas")
#> 
#> MsExpF> fls
#> MsExperimentFiles of length  3 
#> [["mzmls"]] f1.mzML f2.mzML
#> [["mzids"]] id1.mzid
#> [["fasta"]] file.fas
#> 
#> MsExpF> ## A new MsExperimentFiles containing mzML or mzid files
#> MsExpF> fls[1]
#> MsExperimentFiles of length  1 
#> [["mzmls"]] f1.mzML f2.mzML
#> 
#> MsExpF> fls["mzids"]
#> MsExperimentFiles of length  1 
#> [["mzids"]] id1.mzid
#> 
#> MsExpF> ## The actual file names
#> MsExpF> fls[[1]]
#> [1] "/path/to/f1.mzML" "/path/to/f2.mzML"
#> 
#> MsExpF> fls[[2]]
#> [1] "/another/path/to/id1.mzid"
#> 
#> MsExpF> fls[["fasta"]]
#> [1] "file.fas"
#> 
#> MsExpF> ## None of the files used in this example actually exist
#> MsExpF> existMsExperimentFiles(fls)

#> mzmls: 0 out of 2 exist(s)
#> mzids: 0 out of 1 exist(s)
#> fasta: 0 out of 1 exist(s)

experimentFiles(msexp) <- fls
msexp

#> Object of class MsExperiment 
#>  Files: mzmls, mzids, fasta