A class to store experiment files

The `MsExperimentFiles` class stores files that are part of a mass spectrometry experiment. The objects are created with the `MsExperimentFiles()` function.

MsExperimentFiles(..., metadata = list())

Arguments

...
Either a named list or a set of named vectors. All elements are coerced to characters.

metadata
`list()` holding arbitrary R objects as annotations.

Author

Laurent Gatto

Examples

fls <- MsExperimentFiles(mzmls = c("/path/to/f1.mzML", "/path/to/f2.mzML"),
                         mzids = "/another/path/to/id1.mzid",
                         fasta = "file.fas")
fls

#> MsExperimentFiles of length 3
#> [["mzmls"]] /path/to/f1.mzML /path/to/f2.mzML
#> [["mzids"]] /another/path/to/id1.mzid
#> [["fasta"]] file.fas

## A new MsExperimentFiles containing mzML or mzid files
fls[1]

#> MsExperimentFiles of length 1
#> [["mzmls"]] /path/to/f1.mzML /path/to/f2.mzML

fls["mzids"]

#> MsExperimentFiles of length 1
#> [["mzids"]] /another/path/to/id1.mzid

## The actual file names
fls[[1]]

#> [1] "/path/to/f1.mzML" "/path/to/f2.mzML"

fls[[2]]

#> [1] "/another/path/to/id1.mzid"

fls[["fasta"]]

#> [1] "file.fas"